+++ library/renderer.d.ts
@@ -1,0 +1,7 @@
+import type { ReactNativeElectron } from './src/types'
+
+declare global {
+  interface Window {
+    ReactNativeElectron: ReactNativeElectron
+  }
+}