+++ library/src/apis/Platform.ts
@@ -1,6 +1,6 @@
-import type { PlatformSelection, ReactNativeElectron } from '../types.js'
+import type { PlatformSelection } from '../types.js'
 
-const remote = window.ReactNativeElectron as ReactNativeElectron
+const remote = window.ReactNativeElectron
 
 export const isElectron = true
 
@@ -9,8 +9,14 @@ export const OS = 'web'
 export const nativeOS = remote.platform
 
 export const select = (selection: PlatformSelection): unknown => {
-  if (remote && remote.platform in selection) return selection[remote.platform]
-  if ('electron' in selection) return selection.electron
-  if ('web' in selection) return selection.web
+  if (remote && remote.platform in selection) {
+    return selection[remote.platform as keyof PlatformSelection]
+  }
+  if ('electron' in selection) {
+    return selection.electron
+  }
+  if ('web' in selection) {
+    return selection.web
+  }
   return selection.default
 }